+++ src/sockfuncs.h
@@ -42,8 +42,8 @@ enum {
 int sock_connect(const char *addr, short port, int timout_ms);
 int sock_setbufsize(int *s, int send_size, int recv_size);
 int sock_isdisconnected(int *s);
-int sock_send(int *s, const char *buf, int len, int timout_ms);
-int sock_recv(int *s, char *buf, int len, int timout_ms);
+int sock_send(int *s, const char *buf, size_t len, int timout_ms);
+ssize_t sock_recv(int *s, char *buf, int len, int timout_ms);
 int sock_select(int *s, int timout_ms, int mode);
 int sock_nonblock(int *s);
 int sock_block(int *s);
